Added tests for the changes in Alipay+ HK, Kakaopay and TNG. Fix an issue of 'class Omise_Payment_Creditcard' not found when user tries to deactivate WooCommerce.
